ffda-domain-director: check wireless-if status
Devices with ath10k sometimes respond to a scan request (iw dev client0
scan) with "command failed: Resource busy (-16)".

libiwinfo will stall the entire process in this case, never acquiring a
target domain.

We execute a scan via "iw" first. If this returns an error, the radio is
excluded when scanning with iwinfo.