ffda-domain-director: alter disable and config behavior
This commit allows the domain-director settings to be contained in the
domain rather than the global site.

This way migration is possible to be set only for specific domains, most
likely the fallback domain.

Also the user is now given the option to disable the domain-director as
a whole. This means the director

 - won't query the master
 - won't switch domains