ffda-domain-director: scan with private WiFi enabled
Currently, no scan is performed on the radio in case private WiFi is
active on this radio. As private WiFi is enabled on both radios by
default (and non-configurable in the GUI).

In case the node has no neighbours and no geo location set the node will
be directed to te default domain.

The current "fix" is more like a hack due to time-constraints when it
was implemented.

This implements a proper workaround which enables scanning even with
private WiFi enabled.